[update]: about 타입스트립트 업데이트
